+++ lib/src/utilitarios/Parametros.dart
@@ -1,12 +1,17 @@
+import 'package:retorno_sucesso_ou_erro_package/retorno_sucesso_ou_erro_package.dart';
+
 import '../../auth_google_package.dart';
 
-class ParametrosRecuperarSenhaEmail {
+class ParametrosRecuperarSenhaEmail implements ParametrosRetornoResultado {
   final String email;
 
   ParametrosRecuperarSenhaEmail({required this.email});
+
+  @override
+  String get mensagemErro => "Erro ao recuperar a senha pelo e-mail";
 }
 
-class ParametrosSignIn {
+class ParametrosSignIn implements ParametrosRetornoResultado {
   final String? email;
   final String? pass;
   final ResultadoUsuario? user;
@@ -16,4 +21,7 @@ class ParametrosSignIn {
     this.pass,
     this.user,
   });
+
+  @override
+  String get mensagemErro => "Erro ao fazer o SignIn";
 }