+++ CHANGELOG.md
@@ -1,4 +1,9 @@
-## 1.3.0
+## 1.2.1
+
+* TODO: Atualização para retorno_sucesso_ou_erro_package 2.0.2.
+* TODO: Usecase e repositorio substituidos por RetornoResultadoPresenter.
+
+## 1.2.0
 
 * TODO: Ajustes para flutter 2.0.0.